# Conflicts:
#	chd-attence-domain/src/main/java/cn/sshpro/manager/pojo/Student.java
#	chd-attence-web/src/main/java/cn/sshpro/manager/controller/StudentController.java
#	chd-attence-web/src/main/webapp/WEB-INF/vm/layout/header.vm

+++ chd-attence-domain/src/main/java/cn/sshpro/manager/pojo/CourseClass.java
@@ -4,22 +4,22 @@ package cn.sshpro.manager.pojo;
  * Created by liaoxin on 2017/4/10.
  */
 public class CourseClass extends BaseModel{
-    private Integer classId;
-    private Integer courseId;
+    private Long classId;
+    private Long courseId;
 
-    public Integer getClassId() {
+    public Long getClassId() {
         return classId;
     }
 
-    public void setClassId(Integer classId) {
+    public void setClassId(Long classId) {
         this.classId = classId;
     }
 
-    public Integer getCourseId() {
+    public Long getCourseId() {
         return courseId;
     }
 
-    public void setCourseId(Integer courseId) {
+    public void setCourseId(Long courseId) {
         this.courseId = courseId;
     }
 }
